fix(operator): start telemetry before anything imports the Kubernetes client
Static ESM imports are all evaluated before main() runs, so the k8s client had
already bound node:https' request export by the time the SDK started — the
instrumentation wrapped functions nobody was calling and the API-server traffic
this feature exists to trace produced no spans.

The bin is now a telemetry bootstrap: it starts the SDK, then imports the run
itself (moved to main.ts) dynamically. Arming the SDK is not enough on its own
either — HTTP instrumentation patches the CJS exports through the require hook,
which a pure-ESM process never fires for a core module — so the operator pulls
node:http and node:https through it once and re-syncs the builtin ESM facade.
The pre-delete hook subcommand still starts nothing.

Proven in the built image: a real K8sHttp request now emits an
instrumentation-http client span carrying the method and API path, where before
it emitted none.
