+++ docs/designs/agentconnect-org-operator.md
@@ -66,7 +66,8 @@ only if CRD conversion webhooks become necessary.
 
 | Module                        | Role                                                                                                                                                                                                                     | State         |
 | ----------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ | ------------- |
-| `index.ts`                    | Bin entry: telemetry → config → in-cluster client → elector → controller; SIGTERM drains, then flushes spans.                                                                                                            | done          |
+| `index.ts`                    | Bin entry, telemetry only: starts the SDK, then imports `main.ts` — `node:http(s)` bind their exports at import time, so the client must load after the patch.                                                           | done          |
+| `main.ts`                     | The run itself: config → in-cluster client → elector → controller; SIGTERM drains, then flushes spans.                                                                                                                   | done          |
 | `observability.ts`            | The shared OTel bootstrap with the operator's own name and version; off unless the install sets `OTEL_*` (chart: `operator.extraEnv`).                                                                                   | done          |
 | `config.ts`                   | zod env, fail-fast: prefix + tokenreview ClusterRole (required), master template prefix, resync interval, lease name, watch timeout.                                                                                     | done          |
 | `crd/types.ts`                | Group/kind/finalizer/annotation/condition constants + zod spec/status schemas.                                                                                                                                           | done          |
